fix: 多个selector组件公共特性提取放到mixin issue #1924
