+++ src/frontend/devops-pipeline/src/components/AtomFormComponent/SelectInput/index.vue
@@ -42,21 +42,14 @@
 <script>
     import mixins from '../mixins'
     import scrollMixins from './scrollMixins'
+    import selectorMixins from '../selectorMixins'
     import { debounce, isObject } from '@/utils/util'
 
     export default {
         name: 'select-input',
-        mixins: [mixins, scrollMixins],
+        mixins: [mixins, scrollMixins, selectorMixins],
         props: {
             isLoading: Boolean,
-            options: {
-                type: Array,
-                default: []
-            },
-            optionsConf: {
-                type: Object,
-                default: () => ({})
-            },
             preFilter: {
                 type: Object,
                 default: () => ({})
@@ -70,46 +63,16 @@
                 loading: this.isLoading,
                 selectedPointer: 0,
                 selectedGroupPointer: 0,
-                displayName: '',
-                webUrl: WEB_URL_PIRFIX
+                displayName: ''
             }
         },
         computed: {
-            mergedOptionsConf () {
-                return Object.assign({}, {
-                    url: '',
-                    paramId: 'id',
-                    paramName: 'name'
-                }, this.optionsConf)
-            },
             restProps () {
                 const { options, optionsConf, atomvalue, container, ...restProps } = this.$props
                 return restProps
             },
-            hasUrl () {
-                return this.mergedOptionsConf && this.mergedOptionsConf.url && typeof this.mergedOptionsConf.url === 'string'
-            },
             hasGroup () {
                 return this.mergedOptionsConf && this.mergedOptionsConf.hasGroup
-            },
-            urlParamKeys () {
-                if (this.hasUrl) {
-                    const paramKey = this.mergedOptionsConf.url.match(/\{(.*?)\}/g)
-                    return paramKey ? paramKey.map(key => key.replace(/\{(.*?)\}/, '$1')) : []
-                }
-                return []
-            },
-            queryParams () {
-                const { atomValue = {}, $route: { params = {} } } = this
-                return {
-                    ...params,
-                    ...atomValue
-                }
-            },
-            isLackParam () {
-                return this.urlParamKeys.some(key => {
-                    return this.queryParams.hasOwnProperty(key) && (typeof this.queryParams[key] === 'undefined' || this.queryParams[key] === null || this.queryParams[key] === '')
-                })
             },
             filteredList () {
                 const { displayName, value, optionList } = this
@@ -151,12 +114,6 @@
             },
             hasOption () {
                 return Array.isArray(this.filteredList) && this.filteredList.length > 0
-            },
-            addItemUrl () {
-                const { webUrl, urlParse, mergedOptionsConf: { itemTargetUrl }, $route: { params } } = this
-                const originUrl = /^(http|https):\/\//.test(itemTargetUrl) ? itemTargetUrl : webUrl + itemTargetUrl
-
-                return urlParse(originUrl, params)
             }
         },
         watch: {
@@ -197,11 +154,6 @@
             }
         },
         methods: {
-            urlParse (originUrl, query) {
-                /* eslint-disable */
-                return new Function('ctx', `return '${originUrl.replace(/\{(.*?)\}/g, '\'\+ ctx.$1 \+\'')}'`)(query)
-                /* eslint-enable */
-            },
             handleInput (e) {
                 const { name, value } = e.target
                 this.optionListVisible = true
